Clean up and new API
