+++ src/Moryx/Container/Attributes/DependencyRegistrationAttribute.cs
@@ -1,6 +1,7 @@
 // Copyright (c) 2023, Phoenix Contact GmbH & Co. KG
 // Licensed under the Apache License, Version 2.0
 
+using Microsoft.Extensions.DependencyInjection;
 using System;
 
 namespace Moryx.Container
@@ -72,6 +73,6 @@ namespace Moryx.Container
         /// <summary>
         /// Execute further initialization
         /// </summary>
-        void Initialize(IContainer container);
+        void Initialize(IServiceCollection services);
     }
 }
